factor out common IPython code
